requirements.txt

et-xmlfile
ffmpeg
lunardate
numpy
OpenEXR
openpyxl
packaging
pandas
pillow
pyinstaller
pyinstaller-hooks-contrib
pyluach
PyMeeus
PySide6
python-dateutil
shiboken6
workalendar
XlsxWriter
